net/nanocoap: move reply funcs and app resources to nanocoap sock
