+++ sys/include/net/nanocoap.h
@@ -19,8 +19,8 @@
  * provides high and low level interfaces to CoAP options, including Block.
  *
  * nanocoap includes the core structs to store message information. It also
- * provides helper functions for use before sending and after receiving a
- * message, such as coap_parse() to read an incoming message.
+ * also provides functions to support sending and receiving messages, such as
+ * coap_parse() to read an incoming message.
  *
  * ## Application APIs
  *
@@ -245,16 +245,6 @@ typedef struct {
     size_t cur;                     /**< Offset of the generated content    */
     uint8_t *opt;                   /**< Pointer to the placed option       */
 } coap_block_slicer_t;
-
-/**
- * @brief   Global CoAP resource list
- */
-extern const coap_resource_t coap_resources[];
-
-/**
- * @brief   Number of entries in global CoAP resource list
- */
-extern const unsigned coap_resources_numof;
 
 
 /**
@@ -1362,27 +1352,6 @@ static inline size_t coap_put_option_ct(uint8_t *buf, uint16_t lastonum,
  */
 /**@{*/
 /**
- * @brief   Build reply to CoAP block2 request
- *
- * This function can be used to create a reply to a CoAP block2 request
- * packet. In addition to @ref coap_build_reply, this function checks the
- * block2 option and returns an error message to the client if necessary.
- *
- * @param[in]   pkt         packet to reply to
- * @param[in]   code        reply code (e.g., COAP_CODE_204)
- * @param[out]  rbuf        buffer to write reply to
- * @param[in]   rlen        size of @p rbuf
- * @param[in]   payload_len length of payload
- * @param[in]   slicer      slicer to use
- *
- * @returns     size of reply packet on success
- * @returns     <0 on error
- */
-ssize_t coap_block2_build_reply(coap_pkt_t *pkt, unsigned code,
-                                uint8_t *rbuf, unsigned rlen, unsigned payload_len,
-                                coap_block_slicer_t *slicer);
-
-/**
  * @brief   Builds a CoAP header
  *
  * Caller *must* ensure @p hdr can hold the header and the full token!
@@ -1398,65 +1367,6 @@ ssize_t coap_block2_build_reply(coap_pkt_t *pkt, unsigned code,
  */
 ssize_t coap_build_hdr(coap_hdr_t *hdr, unsigned type, uint8_t *token,
                        size_t token_len, unsigned code, uint16_t id);
-
-/**
- * @brief   Build reply to CoAP request
- *
- * This function can be used to create a reply to any CoAP request packet.  It
- * will create the reply packet header based on parameters from the request
- * (e.g., id, token).
- *
- * Passing a non-zero @p payload_len will ensure the payload fits into the
- * buffer along with the header. For this validation, payload_len must include
- * any options, the payload marker, as well as the payload proper.
- *
- * @param[in]   pkt         packet to reply to
- * @param[in]   code        reply code (e.g., COAP_CODE_204)
- * @param[out]  rbuf        buffer to write reply to
- * @param[in]   rlen        size of @p rbuf
- * @param[in]   payload_len length of payload
- *
- * @returns     size of reply packet on success
- * @returns     <0 on error
- * @returns     -ENOSPC if @p rbuf too small
- */
-ssize_t coap_build_reply(coap_pkt_t *pkt, unsigned code,
-                         uint8_t *rbuf, unsigned rlen, unsigned payload_len);
-
-/**
- * @brief   Handle incoming CoAP request
- *
- * This function will find the correct handler, call it and write the reply
- * into @p resp_buf.
- *
- * @param[in]   pkt             pointer to (parsed) CoAP packet
- * @param[out]  resp_buf        buffer for response
- * @param[in]   resp_buf_len    size of response buffer
- *
- * @returns     size of reply packet on success
- * @returns     <0 on error
- */
-ssize_t coap_handle_req(coap_pkt_t *pkt, uint8_t *resp_buf, unsigned resp_buf_len);
-
-/**
- * @brief   Pass a coap request to a matching handler
- *
- * This function will try to find a matching handler in @p resources and call
- * the handler.
- *
- * @param[in]   pkt             pointer to (parsed) CoAP packet
- * @param[out]  resp_buf        buffer for response
- * @param[in]   resp_buf_len    size of response buffer
- * @param[in]   resources       Array of coap endpoint resources
- * @param[in]   resources_numof length of the coap endpoint resources
- *
- * @returns     size of the reply packet on success
- * @returns     <0 on error
- */
-ssize_t coap_tree_handler(coap_pkt_t *pkt, uint8_t *resp_buf,
-                          unsigned resp_buf_len,
-                          const coap_resource_t *resources,
-                          size_t resources_numof);
 
 /**
  * @brief   Convert message code (request method) into a corresponding bit field
@@ -1503,43 +1413,6 @@ int coap_parse(coap_pkt_t *pkt, uint8_t *buf, size_t len);
 void coap_pkt_init(coap_pkt_t *pkt, uint8_t *buf, size_t len, size_t header_len);
 
 /**
- * @brief   Create CoAP reply (convenience function)
- *
- * This is a simple wrapper that allows for building CoAP replies for simple
- * use-cases.
- *
- * The reply will be written to @p buf. If @p payload and @p payload_len are
- * non-zero, the payload will be copied into the resulting reply packet.
- *
- * @param[in]   pkt         packet to reply to
- * @param[in]   code        reply code (e.g., COAP_CODE_204)
- * @param[out]  buf         buffer to write reply to
- * @param[in]   len         size of @p buf
- * @param[in]   ct          content type of payload
- * @param[in]   payload     ptr to payload
- * @param[in]   payload_len length of payload
- *
- * @returns     size of reply packet on success
- * @returns     <0 on error
- * @returns     -ENOSPC if @p buf too small
- */
-ssize_t coap_reply_simple(coap_pkt_t *pkt,
-                          unsigned code,
-                          uint8_t *buf, size_t len,
-                          unsigned ct,
-                          const uint8_t *payload, uint8_t payload_len);
-
-/**
- * @brief   Reference to the default .well-known/core handler defined by the
- *          application
- */
-extern ssize_t coap_well_known_core_default_handler(coap_pkt_t *pkt, \
-                                                    uint8_t *buf, size_t len,
-                                                    void *context);
-/**@}*/
-
-
-/**
  * @brief   Checks if a CoAP resource path matches a given URI
  *
  * Builds on strcmp() with rules specific to URI path matching
@@ -1555,6 +1428,7 @@ extern ssize_t coap_well_known_core_default_handler(coap_pkt_t *pkt, \
  * @return >0 if the resource path sorts after the URI
  */
 int coap_match_path(const coap_resource_t *resource, uint8_t *uri);
+/**@}*/
 
 #if defined(MODULE_GCOAP) || defined(DOXYGEN)
 /**
@@ -1597,16 +1471,6 @@ static inline uint32_t coap_get_observe(coap_pkt_t *pkt)
 }
 /**@}*/
 #endif
-
-/**
- * @brief   Resource definition for the default .well-known/core handler
- */
-#define COAP_WELL_KNOWN_CORE_DEFAULT_HANDLER \
-    { \
-        .path = "/.well-known/core", \
-        .methods = COAP_GET, \
-        .handler = coap_well_known_core_default_handler \
-    }
 
 #ifdef __cplusplus
 }
